added upload_point_id to files table and updated models
A file must always contain a reference to it's upload point. The file type
is in the files table and upload_points table because in the future the
file type might change slightly for a particular upload point, but for
backwards comparibility you would want to keep all the old files that were
uploaded there in their old format (and maybe run a script to convert them
to the new one)

E.g there might be a format relating to a video that is encoded with a
certain codec. In the future we might start using a new codec so we would
create a new file type which would correspond to this new codec and make
this the file type of the upload point. Anything that has already been
uploaded though will still have it's old file_type id so the system knows
what format it is.