Adapted ordererable lists so that automatically scroll.
Had to change to be inline because of issue when dropdown menu was
opened from an ajax select that had been scrolled. It would appear in the
position as if it was not scrolled rendering it useless. Adding position
relative to the search box, which I guess is what's intended, doesn't work
because then the dropdown doesn't overflow outside the orderable list
container as a native drop down would.