View count request no longer made if in admin override mode.
It is now held back until it becomes live for everyone. Any view count
requests when the item is not public are ignored. This means now if admin
mode has previously been enabled, and then the item goes live and the user
watches it, the count will be registered. Only one view count request is
sent during the lifetime of the player.