Fix so that media items not shown if in a playlist where publish time not passed
Previously the playlist would not be shown but the media item files were
still accessible. Also when logged in to the cms with the correct
permission you could still see the playlist page at the url. This was
intented as it means an admin can check everything is working before the
page goes public but changes would need making to make sure that the media
item files aren't accessible to the public and live streams can't be
advertised as live when the public have no way of getting to them.