Header is now checked for session id over cookie
Fixes issue where cookie not set on first request from iframe but is on
the ajax responses, which was occuring in safari. This meant 2 sessions
would be created but the csrf token in the page would be the one linked
with the first one. Now only one session will be created.