Items can now be promoted from the cms, and multiple items may be promoted
If multiple items are promoted all visitors will see the one that was
promoted most recently.
If that is then unpromoted and users haven't seen the one which was
promoted before that one, they will then see that item etc.

This could be made a bit smarter in the future so all promoted items
appear in the carousel and users are shown all promoted videos with the
most recent being shown on the first home page load.